+++ src/main/java/com/hh/attendance/dao/UserMapper.java
@@ -4,5 +4,5 @@ import com.hh.attendance.pojo.User;
 import org.apache.ibatis.annotations.Param;
 
 public interface UserMapper {
-    User getUserById(@Param("userId") Long userId);
+    User getUserById(@Param("userId") Integer userId);
 }
