+++ src/main/java/com/hh/attendance/service/impl/UserService.java
@@ -4,5 +4,5 @@ import com.hh.attendance.pojo.User;
 import org.springframework.web.bind.annotation.RequestParam;
 
 public interface UserService {
-    User getUserById(@RequestParam("userId") Long userId);
+    User getUserById(@RequestParam("userId") Integer userId);
 }
